[R-package] move R source files into R-package, reduce duplication in build_r.R (#3087)
* [R-package] move R source files into R-package

* fix linting warning

* stuff
