+++ R-package/src/R_object_helper.h
